Address review: make the timestamp chain uint32_t
Both reviewers caught that the new sentinel check reintroduces a 2038
problem. LocationProvider::getTimestamp() returned long, which is 32 bit
signed on every target here, so an epoch at or past 2147483648
(19 Jan 2038) reads back as negative and the new 'timestamp > 0' guard
rejects it, blocking time sync for a perfectly valid GPS time.

This was a regression introduced by the guard itself: before it, the
value went straight into setCurrentTime(uint32_t), and the implicit
conversion wrapped the negative long back to the correct unsigned value.

Make the whole chain uint32_t, matching mesh::RTCClock, whose
getCurrentTime() and setCurrentTime() are already uint32_t: the
interface, both overrides, RAK12500LocationProvider::_epoch, and the
local in the consumer, which now compares against 0 with !=.

+++ src/helpers/sensors/LocationProvider.h
@@ -17,7 +17,9 @@ public:
     virtual bool isValid() = 0;
     // UTC seconds since the epoch, or 0 when no valid time is available.
     // Callers must treat 0 as "no time" rather than as a timestamp.
-    virtual long getTimestamp() = 0;
+    // Unsigned to match mesh::RTCClock, and so times past 2038 stay
+    // representable on targets where long is 32 bit signed.
+    virtual uint32_t getTimestamp() = 0;
     virtual void sendSentence(const char * sentence);
     virtual void reset() = 0;
     virtual void begin() = 0;
